refactor: stt 관련 상수 통합 및 이름 변경
- 기존 STT 관련 상수(modules/speeches/speeches.constants.ts) 내용을 공통 상수로 통합
- 공통 상수 파일명 변경(common/constants/speech.constant.ts -> speech.constants.ts)
