--- backend/src/modules/speeches/speeches.constants.ts
@@ -1,10 +1,0 @@
-// 녹음 파일 MIME 타입
-export const allowedMimeTypes = [
-  'audio/wav',
-  'audio/webm',
-  'audio/mpeg',
-  'audio/mp3',
-  'audio/ogg',
-];
-
-export const AUDIOFILE_MAX_SIZE_BYTES = 8 * 1024 * 1024; // 8MB = 3분 wav 파일 약 6MB + 버퍼 2MB
